Split rounding routines
