+++ chainerx_cc/chainerx/cuda/kernel_regist.h
@@ -1,7 +1,14 @@
 #pragma once
 
+#include <cstdint>
+
+#include "chainerx/array.h"
 #include "chainerx/cuda/cuda_backend.h"
+#include "chainerx/device.h"
+#include "chainerx/dtype.h"
 #include "chainerx/kernel_registry.h"
+
+// TODO(niboshi): Rename this file to *.cuh and include cuda/elementwise.cuh.
 
 // Register an kernel statically in CudaBackend.
 #define CHAINERX_CUDA_REGISTER_KERNEL(key_kernel_cls, kernel_cls)                                       \
