+++ chainerx_cc/chainerx/kernels/math.h
@@ -113,20 +113,6 @@ public:
     virtual void Call(Scalar x1, const Array& x2, const Array& out) = 0;
 };
 
-class CeilKernel : public Kernel {
-public:
-    static const char* name() { return "Ceil"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
-class FloorKernel : public Kernel {
-public:
-    static const char* name() { return "Floor"; }
-
-    virtual void Call(const Array& x, const Array& out) = 0;
-};
-
 class IsNanKernel : public Kernel {
 public:
     static const char* name() { return "IsNan"; }
