+++ chainerx_cc/chainerx/routines/math.cc
@@ -503,26 +503,6 @@ Array Power(const Array& x1, Scalar x2) { return internal::Binary(&PowerASImpl, 
 
 Array Power(Scalar x1, const Array& x2) { return internal::Binary(&PowerSAImpl, x1, x2, GetArithmeticResultDtype(x1, x2)); }
 
-Array Ceil(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<CeilKernel>(x, out);
-    }
-    return out;
-}
-
-Array Floor(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<FloorKernel>(x, out);
-    }
-    return out;
-}
-
 Array IsNan(const Array& x) {
     Array out = Empty(x.shape(), Dtype::kBool, x.device());
     {
