+++ chainerx_cc/chainerx/routines/math.h
@@ -87,8 +87,4 @@ Array IsInf(const Array& x);
 
 Array IsFinite(const Array& x);
 
-Array Ceil(const Array& x);
-
-Array Floor(const Array& x);
-
 }  // namespace chainerx
