+++ chainerx_cc/chainerx/routines/rounding.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "chainerx/array.h"
+
+namespace chainerx {
+
+Array Ceil(const Array& x);
+
+Array Floor(const Array& x);
+
+}  // namespace chainerx
